Exclude FileBasedFindingsReport and move methods to print debt into debtSumming
