+++ detekt-cli/src/main/resources/default-detekt-config.yml
@@ -23,7 +23,7 @@ console-reports:
   #  - 'ComplexityReport'
   #  - 'NotificationReport'
   #  - 'FindingsReport'
-  #  - 'FileBasedFindingsReport'
+     - 'FileBasedFindingsReport'
   #  - 'BuildFailureReport'
 
 comments:
