+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/printer/rulesetpage/ConfigPrinter.kt
@@ -93,7 +93,7 @@ object ConfigPrinter : DocumentationPrinter<List<RuleSetPage>> {
         #  - 'ComplexityReport'
         #  - 'NotificationReport'
         #  - 'FindingsReport'
-        #  - 'FileBasedFindingsReport'
+           - 'FileBasedFindingsReport'
         #  - 'BuildFailureReport'
     """.trimIndent()
